Fixed saved states - broke format of course, we need to come up with a better way to do this. Thanks jah :)
svn-id: r3846

+++ scumm.h
@@ -1719,7 +1719,7 @@ struct ScummDebugger {
 };
 
 struct SaveLoadEntry {
-	uint16 offs;
+	uint32 offs;
 	uint8 type;
 	uint8 size;
 };
